engine/requirements.txt
kafka-python
oracledb
requests
pandas
